+++ src/pages/LandingPage.js
@@ -1,0 +1,4 @@
+function LandingPage() {
+  return <h1>랜딩페이지입니다</h1>;
+}
+export default LandingPage;
